feat: 파일 업로드 다운르도 이미지 로드

+++ src/main/java/com/codevumc/codev_backend/domain/CoPhotoOfStudy.java
@@ -16,6 +16,6 @@ public class CoPhotoOfStudy {
     private String fileName;
     private String co_filePath;
     private String co_fileDownloadPath;
-    private Double co_fileSize;
+    private long co_fileSize;
     private Timestamp createdAt;
 }
